added all files to package 'duke'
